+++ spec/javascripts/app/views/publisher_mention_view_spec.js
@@ -19,6 +19,7 @@ describe("app.views.PublisherMention", function() {
       expect(call.args[0].typeaheadInput.selector).toBe("#publisher .typeahead-mention-box");
       expect(call.args[0].customSearch).toBeTruthy();
       expect(call.args[0].autoselect).toBeTruthy();
+      expect(call.args[0].remoteRoute).toBe("/contacts");
     });
 
     it("calls bindTypeaheadEvents", function() {